on the way to work - patch 1

+++ src/preprocess_LibriSpeech.py
@@ -13,7 +13,7 @@ features on the 100 hour Librispeech dataset.
 
 """
 import os
-import glob
+import glob2
 import soundfile as sf
 from python_speech_features import mfcc
 import numpy as np
@@ -35,9 +35,9 @@ def compute_mfcc(audio_data, sample_rate):
 
     audio_data = audio_data - np.mean(audio_data)
     audio_data = audio_data / np.max(audio_data)
-    mfcc_feat = mfcc(audio_data, sample_rate, winlen=0.025, winstep=0.01,
-                     numcep=13, nfilt=26, nfft=512, lowfreq=0, highfreq=None,
-                     preemph=0.97, ceplifter=22, appendEnergy=True)
+    mfcc_feat = mfcc(audio_data, sample_rate, winlen = 0.025, winstep = 0.01,
+                     numcep = 161, nfilt = 322, nfft = 512, lowfreq = 0, highfreq = None,
+                     preemph = 0.97, ceplifter = 22, appendEnergy = True)
     return mfcc_feat
 
 
@@ -66,20 +66,20 @@ def make_example(seq_len, spec_feat, labels):
 
     '''
     # Feature lists for the sequential features of the example
-    feats_list = [tf.train.Feature(float_list=tf.train.FloatList(value=frame))
+    feats_list = [tf.train.Feature(float_list = tf.train.FloatList(value = frame))
                   for frame in spec_feat]
-    feat_dict = {"feats": tf.train.FeatureList(feature=feats_list)}
-    sequence_feats = tf.train.FeatureLists(feature_list=feat_dict)
+    feat_dict = {"feats": tf.train.FeatureList(feature = feats_list)}
+    sequence_feats = tf.train.FeatureLists(feature_list = feat_dict)
 
     # Context features for the entire sequence
-    len_feat = tf.train.Feature(int64_list=tf.train.Int64List(value=[seq_len]))
-    label_feat = tf.train.Feature(int64_list=tf.train.Int64List(value=labels))
+    len_feat = tf.train.Feature(int64_list = tf.train.Int64List(value = [seq_len]))
+    label_feat = tf.train.Feature(int64_list = tf.train.Int64List(value = labels))
 
-    context_feats = tf.train.Features(feature={"seq_len": len_feat,
-                                               "labels": label_feat})
+    context_feats = tf.train.Features(feature = {"seq_len": len_feat,
+                                                 "labels": label_feat})
 
-    ex = tf.train.SequenceExample(context=context_feats,
-                                  feature_lists=sequence_feats)
+    ex = tf.train.SequenceExample(context = context_feats,
+                                  feature_lists = sequence_feats)
 
     return ex.SerializeToString()
 
@@ -103,13 +103,12 @@ def process_data(partition):
     transcripts = {}
     utt_len = {}  # Required for sorting the utterances based on length
 
-    for filename in glob.iglob(partition+'/**/*.txt', recursive=True):
-        with open(filename, 'r') as file:
-            for line in file:
+    for filename in glob2.iglob(partition + '/**/*.txt'):
+        with open(filename, 'r') as f:
+            for line in f:
                 parts = line.split()
                 audio_file = parts[0]
-                file_path = os.path.join(os.path.dirname(filename),
-                                         audio_file+'.flac')
+                file_path = os.path.join(os.path.dirname(filename), audio_file + '.flac')
                 audio, sample_rate = sf.read(file_path)
                 feats[audio_file] = compute_mfcc(audio, sample_rate)
                 utt_len[audio_file] = feats[audio_file].shape[0]
@@ -124,16 +123,18 @@ def create_records():
     into integers, and generates sequence examples for each utterance.
     Multiple sequence records are then written into TFRecord files.
     """
-    for partition in sorted(glob.glob(AUDIO_PATH+'/*')):
+    for partition in sorted(glob2.glob(AUDIO_PATH + '/*')):
+        if os.path.isfile(partition):
+            continue
         print('Processing' + partition)
         feats, transcripts, utt_len = process_data(partition)
-        sorted_utts = sorted(utt_len, key=utt_len.get)
+        sorted_utts = sorted(utt_len, key = utt_len.get)
         # bin into groups of 100 frames.
-        max_t = int(utt_len[sorted_utts[-1]]/100)
-        min_t = int(utt_len[sorted_utts[0]]/100)
+        max_t = int(utt_len[sorted_utts[-1]] / 100)
+        min_t = int(utt_len[sorted_utts[0]] / 100)
 
         # Create destination directory
-        write_dir = '../data/librispeech/processed/' + partition.split('/')[-1]
+        write_dir = os.path.join(AUDIO_PATH, 'processed', partition.split('/')[-1])
         if tf.gfile.Exists(write_dir):
             tf.gfile.DeleteRecursively(write_dir)
         tf.gfile.MakeDirs(write_dir)
@@ -143,44 +144,39 @@ def create_records():
             writer = {}
             count = {}
             print('Processing training files...')
-            for i in range(min_t, max_t+1):
-                filename = os.path.join(write_dir, 'train' + '_' + str(i) +
-                                        '.tfrecords')
+            for i in range(min_t, max_t + 1):
+                filename = os.path.join(write_dir, 'train' + '_' + str(i) + '.tfrecords')
                 writer[i] = tf.python_io.TFRecordWriter(filename)
                 count[i] = 0
 
             for utt in tqdm(sorted_utts):
-                example = make_example(utt_len[utt], feats[utt].tolist(),
-                                       transcripts[utt])
-                index = int(utt_len[utt]/100)
+                example = make_example(utt_len[utt], feats[utt].tolist(), transcripts[utt])
+                index = int(utt_len[utt] / 100)
                 writer[index].write(example)
                 count[index] += 1
 
-            for i in range(min_t, max_t+1):
+            for i in range(min_t, max_t + 1):
                 writer[i].close()
             print(count)
 
             # Remove bins which have fewer than 20 utterances
-            for i in range(min_t, max_t+1):
+            for i in range(min_t, max_t + 1):
                 if count[i] < 20:
-                    os.remove(os.path.join(write_dir, 'train' +
-                                           '_' + str(i) + '.tfrecords'))
+                    os.remove(os.path.join(write_dir, 'train' + '_' + str(i) + '.tfrecords'))
         else:
             # Create single TFRecord for dev and test partition
-            filename = os.path.join(write_dir, os.path.basename(write_dir) +
-                                    '.tfrecords')
+            filename = os.path.join(write_dir, os.path.basename(write_dir) + '.tfrecords')
             print('Creating', filename)
             record_writer = tf.python_io.TFRecordWriter(filename)
             for utt in sorted_utts:
-                example = make_example(utt_len[utt], feats[utt].tolist(),
-                                       transcripts[utt])
+                example = make_example(utt_len[utt], feats[utt].tolist(), transcripts[utt])
                 record_writer.write(example)
             record_writer.close()
-            print('Processed '+str(len(sorted_utts))+' audio files')
+            print('Processed ' + str(len(sorted_utts)) + ' audio files')
 
 # Audio path is the location of the directory that contains the librispeech
 # data partitioned into three folders: dev-clean, train-clean-100, test-clean
-AUDIO_PATH = '../data/librispeech/audio/'
+AUDIO_PATH = '../data/LibriSpeech'
 ALPHABET = "ABCDEFGHIJKLMNOPQRSTUVWXYZ' "
 CHAR_TO_IX = {ch: i for (i, ch) in enumerate(ALPHABET)}
